выделил сцену для диаграмм в пакет
