+++ ii_constructor/packages/diagrammscene/README.md
@@ -1,0 +1,1 @@
+Обёртка над графической сценой Qt для представления диаграмм
