+++ ii_constructor/packages/diagrammscene/requirements.txt
@@ -1,0 +1,1 @@
+pyside6
