+++ requirements.txt
@@ -4,3 +4,4 @@
 -e ./ii_constructor/packages/mysqlrepo
 -e ./ii_constructor/packages/inmemoryrepo
 -e ./ii_constructor/packages/gui
+-e ./ii_constructor/packages/diagrammscene
